% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leastSquares.R
\name{least.squares.estimator}
\alias{least.squares.estimator}
\title{least.squares.estimator}
\usage{
least.squares.estimator(
  data,
  distribution = c("exponential", "shift.exponential", "weibull", "shift.weibull",
    "lognormal", "birnbaum.saunders"),
  plot.fit = FALSE
)
}
\arguments{
\item{data}{the data to estimate parameters from}

\item{distribution}{the distribution to estimate parameters for}

\item{plot.fit}{set to TRUE to plot the least squares fit}
}
\value{
returns a named numeric with the parameters, MTTF, slope, intercept and R.squared
}
\description{
uses least squares to estimate the parameters of a distribution from data
}
