% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximumLikelihood.R
\name{maximum.likelihood.estimator}
\alias{maximum.likelihood.estimator}
\title{maximum.likelihood.estimator}
\usage{
maximum.likelihood.estimator(
  data,
  distribution = c("exponential", "weibull", "lognormal"),
  lower.bound = 0,
  upper.bound = 100,
  tol = 1e-10
)
}
\arguments{
\item{data}{the data to estimate parameters from}

\item{distribution}{the distribution to estimate parameters for}

\item{lower.bound}{the lower bound for numerical estimation for parameters
with no closed form}

\item{upper.bound}{the upper bound for numerical estimation for parameters
with no closed form}

\item{tol}{the desired accuracy for numerical estimation for parameters
with no closed form}
}
\value{
returns a named numeric with the parameters
}
\description{
uses maximum likelihood to estimate the parameters of a distribution from data
}
