% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodOfMoments.R
\name{mom.estimator}
\alias{mom.estimator}
\title{mom.estimator}
\usage{
mom.estimator(
  data,
  distribution = c("exponential", "shifted.exponential", "weibull"),
  lower.bound = 1,
  upper.bound = 100,
  tol = 0.01
)
}
\arguments{
\item{data}{the data to estimate parameters from}

\item{distribution}{the distribution to estimate parameters for}

\item{lower.bound}{the lower bound for numerical estimation for parameters
with no closed form}

\item{upper.bound}{the upper bound for numerical estimation for parameters
with no closed form}

\item{tol}{the desired accuracy for numerical estimation for parameters
with no closed form}
}
\value{
returns a named numeric with the parameters
}
\description{
uses method of moments to estimate the parameters of a distribution from data
}
